git-svn-id: http://svn.boost.org/svn/boost/branches/release@54064 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ libs/graph_parallel/example/breadth_first_search.cpp
@@ -83,7 +83,7 @@ int main(int argc, char* argv[])
     outfile = argv[2];
   else {
     outfile = filename;
-    int i = outfile.rfind('.');
+    size_t i = outfile.rfind('.');
     if (i != std::string::npos)
       outfile.erase(outfile.begin() + i, outfile.end());
     outfile += "-bfs.dot";
